refactor: rename module name of arcade-info list
