+++ src/features/arcade-info/list.tsx
